flogi/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.2.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==4.2.0
bleach==5.0.1
boto3==1.29.5
botocore==1.32.5
celery==5.3.5
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==4.0.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
Django==4.1.4
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-yasg==1.21.4
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.16.2
frozenlist==1.4.0
googlemaps==4.10.0
hyperlink==21.0.0
idna==3.4
imageio==2.33.0
imageio-ffmpeg==0.4.9
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kombu==5.3.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
moviepy==1.0.3
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.3
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.0
openai==0.27.8
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221213
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pluggy==1.2.0
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.41
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.2
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.4.2
qtconsole==5.4.0
QtPy==2.2.1
random-username==1.0.2
redis==5.0.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.7.0
Send2Trash==1.8.0
service-identity==21.1.0
sib-api-v3-sdk==7.6.0
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.0
terminado==0.17.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
Twisted==22.10.0
txaio==23.1.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
vine==5.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
yarl==1.9.2
zope.interface==6.0
